refactor: replace action.ts to actions.ts
